refactor: separate env variables for dev and prod
