+++ .github/workflows/tests.yml
@@ -11,14 +11,14 @@ jobs:
 
     strategy:
       matrix:
-        node-version: [16.x, 18.x]
+        node-version: [18.x, 19.x, 20.x]
 
     steps:
     - name: Checkout
       uses: actions/checkout@v1
 
     - name: Start containers
-      run: docker-compose -f "docker-compose-test.yml" up -d --build
+      run: yarn test:dockerstart
 
     - name: Install node
       uses: actions/setup-node@v3
@@ -34,4 +34,4 @@ jobs:
 
     - name: Stop containers
       if: always()
-      run: docker-compose -f "docker-compose-test.yml" down
+      run: yarn test:dockerstop
