--- .jest/setEnvVars.ts
@@ -1,4 +1,0 @@
-process.env.NODE_ENV = 'dev';
-process.env.TEST_ENABLED = "true";
-process.env.FIRESTORE_EMULATOR_HOST = "localhost:54321"
-process.env.FIRESTORE_PROJECT_ID = "find-a-doc-japan"
